dojson: add urls2marc rule to hepnames
